+++ scripts/make_app.sh
@@ -61,8 +61,8 @@ cat > "$APP/Contents/Info.plist" <<PLIST
     <key>CFBundleName</key><string>$APP_NAME</string>
     <key>CFBundleDisplayName</key><string>$APP_NAME</string>
     <key>CFBundleIdentifier</key><string>$BUNDLE_ID</string>
-    <key>CFBundleVersion</key><string>2.2.14</string>
-    <key>CFBundleShortVersionString</key><string>2.2.14</string>
+    <key>CFBundleVersion</key><string>2.2.15</string>
+    <key>CFBundleShortVersionString</key><string>2.2.15</string>
     <key>CFBundlePackageType</key><string>APPL</string>
     <key>CFBundleIconFile</key><string>AppIcon</string>
     <key>LSMinimumSystemVersion</key><string>13.0</string>